feat: some of blog preview component created
