+++ bootcamp-milestone-2/src/app/components/blogPreview.module.css
@@ -1,0 +1,24 @@
+/* * {
+  padding: 0;
+  margin: 0;
+} */
+
+.pageTitle {
+  margin: 3rem 3rem;
+  text-align: center;
+  background-color: white;
+  opacity: 0.85;
+  box-shadow: 10px 10px 10px black;
+  box-sizing: 50%;
+  border: 4px solid black;
+}
+
+.postDate {
+  margin: 3rem 3rem;
+  text-align: center;
+  background-color: white;
+  opacity: 0.85;
+  box-shadow: 10px 10px 10px black;
+  box-sizing: 50%;
+  border: 4px solid black;
+}
