+++ bootcamp-milestone-2/src/app/contact/contact.module.css
@@ -1,10 +1,3 @@
-input,
-label {
-  display: block;
-  margin: auto;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
 #name,
 #email {
   height: 20px;
@@ -16,7 +9,3 @@ label {
   margin: 0 auto;
 }
 
-label {
-  font-weight: bold;
-  text-align: center;
-}
